requirements.txt
configparser
pymysql
paho-mqtt
